pyproject.toml
[project]
name = "mail-bot"
version = "0.1.0"
description = "Telegram бот для пересылки писем из Яндекс.Почты"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aioimaplib>=2.0.1",
    "fastapi>=0.122.0",
    "flask>=3.0.0",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]
